Organize icons into the correct directory
