+++ src/login/index.html
@@ -35,7 +35,7 @@
         </div>
         <div class="panels">
           <div id="mobile-header">
-            <img src="/static/img/y-logo-small.svg">
+            <img src="/static/img/ylitse-logo-tiny.svg">
             <h1 localization-key="shortTitle"></h1>
           </div>
           <div class="left panel">
@@ -69,7 +69,7 @@
           </div>
           <div class="right panel">
             <div id="right-text">
-              <img src="/static/img/y-logo-large.svg">
+              <img src="/static/img/ylitse-logo-small.svg">
               <h1 localization-key="longTitle"></h1>
               <p localization-key="appDescription"></p>
               <a id="register-button"
