+++ src/register/styles.css
@@ -5,7 +5,7 @@ body {
       rgba(1, 165, 236, 0.58) 0%,
       rgba(89, 58, 214, 0.58) 100%
     ),
-    url(/static/img/milad-fakurian-E8Ufcyxz514-unsplash.jpg) no-repeat fixed
+    url(/static/img/splash-background.jpg) no-repeat fixed
       right top;
   -webkit-background-size: cover;
   -moz-background-size: cover;
